refactor: extract timodoro list helpers
Tests: node --test tests/ui/workspaces/timodoroWorkspacePresenter.test.js

Manual testing: not run (CLI environment)
